Move kernel implementations outside of lookup table utility functions. (#8306)
